Refactoring topics grid
